refactor(DEV-11981): unify Reminder fixtures and handlers
