--- packages/sdk-react/src/mocks/overdueReminders/index.ts
@@ -1,2 +1,0 @@
-export * from './overdueRemindersHandlers';
-export * from './overdueRemindersFixtures';
