--- packages/sdk-react/src/mocks/paymentReminders/index.ts
@@ -1,4 +1,0 @@
-export * from './paymentRemindersFixtures';
-export * from './paymentRemindersHandlers';
-export * from './reminderHandlers';
-export * from './reminderFixtures';
