+++ packages/sdk-react/src/mocks/reminders/index.ts
@@ -1,0 +1,1 @@
+export { remindersHandlers } from '@/mocks/reminders/remindersHandlers';
